% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{errors}
\alias{errors}
\title{Errors of a list of formulas.}
\usage{
errors(endogenous, exogenous, formulas)
}
\arguments{
\item{endogenous}{A list of variables to be used in evaluation (variables may be scalars, vectors or matrices)}

\item{exogenous}{A list of parameters to be used in evaluation}

\item{formulas}{A list of formulas}
}
\value{
A list of real numbers on the same skeleton as the endogenous variables
}
\description{
\code{errors} returns all differences between the left and right sides of a list of formula (errors).
}
\examples{
errors(formulas =
          list(price = price ~ 0.3 * variablecost + coefficient1,
               variablecost  = variablecost ~ 0.5 * price + coefficient2),
       endogenous = list(price = 1, variablecost=0.4),
       exogenous = list(coefficient1=0.56, coefficient2=0.7))
}
