% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSolver.r
\name{presolveSystem}
\alias{presolveSystem}
\title{Pre solve a system of formulas.}
\usage{
presolveSystem(formulas, constants = list())
}
\arguments{
\item{constants}{(optional) A list of coefficients}
}
\value{
A list of formulas, constants and functions
}
\description{
\code{presolveSystem} returns a set of remaining formulas and implied constants and functions
}
\examples{
presolveSystem(formulas =
          list(price = 1 ~ price + coefficient1,
               variablecost  = variablecost ~ 0.5 * price + coefficient2), constants=list(coefficient1=3))
}
