% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\docType{package}
\name{solver}
\alias{solver}
\alias{solver-package}
\title{solver: A package solving a system of equations.}
\description{
The solver package provides two important functions: rsolve (which solves a system) and errors (which evaluates error terms)
}
