% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solveSystem}
\alias{solveSystem}
\title{Solve a system of equations.}
\usage{
solveSystem(formulas, endogenous, exogenous, lowerBounds = NULL,
  upperBounds = NULL, maxIterations = 100, maxError = 5e-09)
}
\arguments{
\item{formulas}{A list of formulas}

\item{endogenous}{A list of variables to be used in evaluation (variables may be scalars, vectors or matrices)}

\item{exogenous}{A list of parameters to be used in evaluation}

\item{lowerBounds}{A list of lower bounds for endogenous variables}

\item{upperBounds}{A list of upper bounds for endogenous variables}

\item{maxIterations}{A maximum number of times the Newton algorithm should be applied (default value = 100)}

\item{maxError}{Maximum total absolute error of all equations (default value = 5e-9)}
}
\value{
A vector of real numbers
}
\description{
\code{solveSystem} returns a set of endogenous variables that solve the system.
}
\examples{
solveSystem(formulas =
          list(price = price ~ 0.3 * variablecost + coefficient1,
               variablecost  = variablecost ~ 0.5 * price + coefficient2),
       endogenous = list(price = 1, variablecost=0.4),
       exogenous = list(coefficient1=0.56, coefficient2=0.7))
solveSystem(formulas =
          list(price = price ~ 0.3 * variablecost + coefficient1,
               variablecost  = variablecost ~ 0.5 * price + coefficient2),
       endogenous = list(price = c(wheat=1, rice=4), variablecost=c(wheat=0.4, rice=0.2)),
       exogenous = list(coefficient1=c(wheat=0.56, rice=4), coefficient2=c(wehat=0.7, rice=0.9)))
}
